fpga robots: Add a library to run robots on FPGA
This includes adding a shell script to make running of robots on
FPGA possible as well.

Unfortunately, the only way to run these robots against the FPGAs
is to use this shell script, because Robot does not allow for
dynamic changes to test suites in the Settings section of a suite.

Change-Id: I96acb0ae1b475e1ecc204c96627aba95efa7acc8